+++ ForgedInvariant/TSCSyncer.hpp
@@ -3,18 +3,19 @@
 
 #pragma once
 #include <Headers/kern_patcher.hpp>
+#include <IOKit/IOTimerEventSource.h>
 
-class TSCSyncer {
-    _Atomic(bool) systemAwake;
-    _Atomic(bool) synchronising;
-    _Atomic(bool) synchronised;
-    _Atomic(int) threadsEngaged;
-    _Atomic(UInt64) targetTSC;
+class TSCForger {
+    atomic_bool systemAwake = ATOMIC_VAR_INIT(true);
+    atomic_bool synchronising = ATOMIC_VAR_INIT(false);
+    atomic_bool synchronised = ATOMIC_VAR_INIT(false);
+    atomic_int threadsEngaged = ATOMIC_VAR_INIT(0);
+    atomic_ullong targetTSC = ATOMIC_VAR_INIT(0);
 
     union {
         struct {
-            UInt8 tscAdjust : 1;
-            UInt8 amd17h : 1;
+            bool tscAdjust : 1;
+            bool amd17h : 1;
             UInt8 _rsvd : 6;
         };
         UInt8 raw {0};
@@ -24,10 +25,11 @@ class TSCSyncer {
     mach_vm_address_t orgXcpmUrgency {0};
     mach_vm_address_t orgTracePoint {0};
     mach_vm_address_t orgClockGetCalendarMicrotime {0};
+    IOTimerEventSource *timer {nullptr};
 
     static void resetAdjust(void *);
     void lockFreq();
-    static void setTscValue(void *);
+    static void sync(void *);
 
     static void wrapXcpmUrgency(int urgency, UInt64 rtPeriod, UInt64 rtDeadline);
     static void wrapTracePoint(void *that, UInt8 point);
@@ -35,10 +37,15 @@ class TSCSyncer {
 
     void processPatcher(KernelPatcher &patcher);
 
+    void startTimer();
+    void stopTimer();
+
+    static void timerAction(OSObject *owner, IOTimerEventSource *timer);
+
+    void syncAll(bool timer = false);
+
     public:
-    static TSCSyncer &singleton();
+    static TSCForger &singleton();
 
     void init();
-    void sync(bool timer = false);
-    bool periodicSyncRequired() const;
 };